+++ xcache-common/src/main/java/com/igeeksky/xcache/props/StatProps.java
@@ -9,21 +9,20 @@ package com.igeeksky.xcache.props;
 public class StatProps {
 
     /**
-     * 缓存统计的时间间隔（可为空）
+     * 缓存统计的时间间隔
      */
     private Long period;
 
     /**
-     * 缓存统计的时间间隔（可为空）
+     * 缓存统计的时间间隔
      * <p>
-     * 默认值：60000 单位：毫秒
+     * 默认值：60000 单位：毫秒 <br>
+     * {@link CacheConstants#DEFAULT_STAT_PERIOD}
      * <p>
      * 如果采用内嵌的缓存统计（写入日志），可以通过此配置调整统计周期。
      * 打印日志由 {@code com.igeeksky.xcache.extension.stat.LogStatMessagePublisher} 完成，日志级别为 INFO。<p>
      * 用户可以通过调整该类的日志配置，以控制缓存统计日志是否输出及输出方式。
      * 譬如输出到独立的统计日志文件，或者通过插件输出到 MQ，然后采集计算实现全局统计。
-     * <p>
-     * {@link CacheConstants#DEFAULT_STAT_PERIOD}
      *
      * @return Long – 缓存统计的时间间隔
      */
