+++ xcache-redis/xcache-redis-common/src/main/java/com/igeeksky/redis/RedisOperator.java
@@ -8,26 +8,75 @@ import java.util.List;
 import java.util.Map;
 
 /**
+ * Redis 客户端操作接口
+ *
  * @author Patrick.Lau
  * @since 0.0.4 2023-10-01
  */
 public interface RedisOperator extends AutoCloseable {
 
+    /**
+     * Redis 服务端正确执行命令时的返回值
+     */
     String OK = "OK";
 
+    /**
+     * 是否为集群模式
+     *
+     * @return true：集群模式；false：单机模式
+     */
     boolean isCluster();
 
+    /**
+     * Redis 客户端是否未关闭连接
+     *
+     * @return true：已打开；false：未打开
+     */
     boolean isOpen();
 
     // String Command --start--
+
+    /**
+     * 获取指定键的值。
+     *
+     * @param key 键
+     * @return 值
+     */
     byte[] get(byte[] key);
 
+    /**
+     * 批量获取指定键的值。
+     *
+     * @param keys 键集
+     * @return 值集
+     */
     List<KeyValue<byte[], byte[]>> mget(byte[]... keys);
 
+    /**
+     * 存储键值对。
+     *
+     * @param key   键
+     * @param value 值
+     * @return result
+     */
     String set(byte[] key, byte[] value);
 
+    /**
+     * 存储键值对，并设置过期时间。
+     *
+     * @param key          键
+     * @param milliseconds 过期时间，单位：毫秒
+     * @param value        值
+     * @return result
+     */
     String psetex(byte[] key, long milliseconds, byte[] value);
 
+    /**
+     * 批量存储键值对。
+     *
+     * @param keyValues 键值对集合
+     * @return result
+     */
     String mset(Map<byte[], byte[]> keyValues);
 
     /**
@@ -49,10 +98,31 @@ public interface RedisOperator extends AutoCloseable {
     // String Command --end--
 
     // Hash Command --start--
+
+    /**
+     * 获取指定键的指定字段的值。
+     *
+     * @param key   键
+     * @param field 字段
+     * @return 值
+     */
     byte[] hget(byte[] key, byte[] field);
 
+    /**
+     * 获取指定键的所有字段的值。
+     *
+     * @param key 键
+     * @return 所有字段的值
+     */
     Map<byte[], byte[]> hgetall(byte[] key);
 
+    /**
+     * 批量获取指定键的指定字段的值。
+     *
+     * @param key   键
+     * @param field 字段集
+     * @return 结果集
+     */
     List<KeyValue<byte[], byte[]>> hmget(byte[] key, byte[]... field);
 
     /**
@@ -62,22 +132,50 @@ public interface RedisOperator extends AutoCloseable {
      */
     List<KeyValue<byte[], byte[]>> hmget(Map<byte[], List<byte[]>> keyFields, int totalSize);
 
+    /**
+     * 设置指定键的指定字段的值。
+     *
+     * @param key   键
+     * @param field 字段
+     * @param value 值
+     * @return true：设置成功；false：设置失败
+     */
     Boolean hset(byte[] key, byte[] field, byte[] value);
 
-    String hmset(byte[] key, Map<byte[], byte[]> map);
+    /**
+     * 批量设置指定键的指定字段的值。
+     *
+     * @param key         键
+     * @param fieldValues 字段 : 值 的集合
+     * @return 命令执行结果：如正确执行，返回 OK。
+     */
+    String hmset(byte[] key, Map<byte[], byte[]> fieldValues);
 
+    /**
+     * 批量设置指定键的指定字段的值。
+     *
+     * @param keyFieldValues 键 : 字段 : 值 的集合
+     * @return 命令执行结果：如正确执行，返回 OK。
+     */
     String hmset(Map<byte[], Map<byte[], byte[]>> keyFieldValues);
 
+    /**
+     * 删除指定键的指定字段。
+     *
+     * @param key    键
+     * @param fields 字段列表
+     * @return 删除的字段数量
+     */
     long hdel(byte[] key, byte[]... fields);
 
+    /**
+     * 批量删除指定键的指定字段。
+     *
+     * @param keyFields 键 : 字段列表
+     * @return 删除的字段数量
+     */
     long hdel(Map<byte[], List<byte[]>> keyFields);
     // Hash Command --end--
-
-    List<byte[]> keys(byte[] matches);
-
-    long clear(byte[] matches);
-
-    void publish(byte[] channel, byte[] message);
 
 
     // Stream Command --start--
@@ -104,26 +202,133 @@ public interface RedisOperator extends AutoCloseable {
 
 
     // Script Command --start--
+
+    /**
+     * 检查给定的脚本是否已经被加载。
+     *
+     * @param digests 脚本的 SHA1 校验和。
+     * @return 一个布尔值列表，指示给定的脚本是否已经被加载。
+     */
     List<Boolean> scriptExists(String... digests);
 
+    /**
+     * 将脚本加载到 Redis 服务器，并返回 SHA1 校验和。
+     *
+     * @param script 脚本对象。
+     * @return 返回 SHA1 校验和。
+     */
     String scriptLoad(RedisScript<?> script);
 
+    /**
+     * 使用指定的键和参数执行 Lua 脚本。
+     *
+     * @param script   脚本对象。
+     * @param keyCount 键的数量。
+     * @param params   脚本的参数。
+     * @return 返回执行脚本的结果。
+     */
     <T> T eval(RedisScript<T> script, int keyCount, byte[]... params);
 
+    /**
+     * 使用指定的键和参数执行 Lua 脚本。
+     *
+     * @param script 脚本对象。
+     * @param keys   键的列表。
+     * @param args   脚本的参数。
+     * @return 返回执行脚本的结果。
+     */
     <T> T eval(RedisScript<T> script, byte[][] keys, byte[]... args);
 
+    /**
+     * 使用指定的键和参数执行只读的 Lua 脚本。
+     *
+     * @param script   脚本对象。
+     * @param keyCount 键的数量。
+     * @param params   脚本的参数。
+     * @return 返回执行脚本的结果。
+     */
     <T> T evalReadOnly(RedisScript<T> script, int keyCount, byte[]... params);
 
+    /**
+     * 使用指定的键和参数执行只读的 Lua 脚本。
+     *
+     * @param script 脚本对象。
+     * @param keys   键的列表。
+     * @param args   脚本的参数。
+     * @return 返回执行脚本的结果。
+     */
     <T> T evalReadOnly(RedisScript<T> script, byte[][] keys, byte[]... args);
 
+    /**
+     * 使用指定的键和参数执行只读的 Lua 脚本。
+     *
+     * @param script   脚本对象
+     * @param keyCount 键数量
+     * @param params   脚本的参数
+     * @return 返回执行脚本的结果
+     */
     <T> T evalsha(RedisScript<T> script, int keyCount, byte[]... params);
 
+    /**
+     * 使用指定的键和参数执行只读的 Lua 脚本。
+     *
+     * @param script 脚本对象。
+     * @param keys   键的列表。
+     * @param args   脚本的参数。
+     * @return 返回执行脚本的结果。
+     */
     <T> T evalsha(RedisScript<T> script, byte[][] keys, byte[]... args);
 
+    /**
+     * 使用指定的键和参数执行只读的 Lua 脚本。
+     *
+     * @param script   脚本对象。
+     * @param keyCount 键的数量。
+     * @param params   脚本的参数。
+     * @return 返回执行脚本的结果。
+     */
     <T> T evalshaReadOnly(RedisScript<T> script, int keyCount, byte[]... params);
 
+    /**
+     * 使用指定的键和参数执行只读的 Lua 脚本。
+     *
+     * @param script 脚本对象。
+     * @param keys   键的列表。
+     * @param args   脚本的参数。
+     * @return 返回执行脚本的结果。
+     */
     <T> T evalshaReadOnly(RedisScript<T> script, byte[][] keys, byte[]... args);
     // Script Command --end--
+
+
+    // BaseCommand --start--
+
+    /**
+     * 获取匹配指定模式的所有键。
+     *
+     * @param matches 匹配模式
+     * @return 键的列表
+     */
+    List<byte[]> keys(byte[] matches);
+
+    /**
+     * 清除匹配指定模式的所有键。
+     *
+     * @param matches 匹配模式
+     * @return 删除的键的数量
+     */
+    long clear(byte[] matches);
+
+    /**
+     * 发布消息。
+     * <p>
+     * 发布订阅模式，用于在多个客户端之间广播消息。
+     *
+     * @param channel 发布消息的频道。
+     * @param message 发布的消息。
+     * @return 收到此消息的订阅者数量
+     */
+    long publish(byte[] channel, byte[] message);
 
     /**
      * 获取 Redis 服务器的当前时间戳。
@@ -131,4 +336,5 @@ public interface RedisOperator extends AutoCloseable {
      * @return UNIX时间戳，毫秒
      */
     long timeMillis();
+    // BaseCommand --end--
 }
