+++ xcache-redis/xcache-redis-core/src/main/java/com/igeeksky/xcache/redis/sync/RedisCacheSyncMessageCodec.java
@@ -4,13 +4,12 @@ package com.igeeksky.xcache.redis.sync;
 import com.igeeksky.xcache.extension.sync.CacheSyncMessage;
 import com.igeeksky.xcache.redis.StreamMessageCodec;
 import com.igeeksky.xtool.core.collection.CollectionUtils;
+import com.igeeksky.xtool.core.collection.Maps;
 import com.igeeksky.xtool.core.lang.ArrayUtils;
 import com.igeeksky.xtool.core.lang.ByteArray;
 import com.igeeksky.xtool.core.lang.codec.Codec;
 import com.igeeksky.xtool.core.lang.codec.StringCodec;
 
-import java.nio.charset.Charset;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
@@ -50,7 +49,7 @@ public class RedisCacheSyncMessageCodec implements StreamMessageCodec<CacheSyncM
     }
 
     public Map<byte[], byte[]> encodeMsg(CacheSyncMessage message) {
-        Map<byte[], byte[]> body = HashMap.newHashMap(4);
+        Map<byte[], byte[]> body = Maps.newHashMap(4);
         body.put(sid.getValue(), stringCodec.encode(message.getSid()));
         body.put(type.getValue(), stringCodec.encode(Integer.toString(message.getType())));
         Set<String> keys = message.getKeys();
@@ -61,7 +60,7 @@ public class RedisCacheSyncMessageCodec implements StreamMessageCodec<CacheSyncM
     }
 
     public CacheSyncMessage decodeMsg(Map<byte[], byte[]> body) {
-        Map<ByteArray, byte[]> temp = HashMap.newHashMap(body.size());
+        Map<ByteArray, byte[]> temp = Maps.newHashMap(body.size());
         body.forEach((k, v) -> temp.put(ByteArray.of(k), v));
 
         CacheSyncMessage message = new CacheSyncMessage();
