+++ xcache-redis/xcache-redis-jedis/src/main/java/com/igeeksky/redis/jedis/AbstractJedisOperator.java
@@ -125,8 +125,8 @@ public abstract class AbstractJedisOperator implements RedisOperator {
     }
 
     @Override
-    public void publish(byte[] channel, byte[] message) {
-
+    public long publish(byte[] channel, byte[] message) {
+        return jedis.publish(channel, message);
     }
 
     @Override
