+++ xcache-redis/xcache-redis-lettuce/src/main/java/com/igeeksky/redis/lettuce/AbstractLettuceFactory.java
@@ -9,15 +9,24 @@ import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
 /**
+ * Lettuce 抽象工厂
+ *
  * @author Patrick.Lau
  * @since 1.0.0 2024/7/21
  */
-public abstract class AbstractLettuceFactory implements RedisOperatorFactory {
+public sealed abstract class AbstractLettuceFactory implements RedisOperatorFactory
+        permits LettuceStandaloneFactory, LettuceSentinelFactory, LettuceClusterFactory {
 
     private final Lock lock = new ReentrantLock();
 
     private volatile RedisStreamOperator streamOperator;
     private volatile AbstractLettuceOperator lettuceOperator;
+
+    /**
+     * 默认构造函数
+     */
+    protected AbstractLettuceFactory() {
+    }
 
     @Override
     public RedisOperator getRedisOperator() {
@@ -49,8 +58,18 @@ public abstract class AbstractLettuceFactory implements RedisOperatorFactory {
         return this.streamOperator;
     }
 
+    /**
+     * 创建 LettuceOperator
+     *
+     * @return {@link AbstractLettuceOperator} Lettuce客户端
+     */
     protected abstract AbstractLettuceOperator createOperator();
 
+    /**
+     * 创建 LettuceStreamOperator
+     *
+     * @return {@link LettuceStreamOperator} Lettuce客户端
+     */
     protected abstract LettuceStreamOperator createStreamOperator();
 
     @Override
