+++ xcache-redis/xcache-redis-lettuce/src/main/java/com/igeeksky/redis/lettuce/LettuceHelper.java
@@ -15,6 +15,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
+ * Lettuce 辅助工具
+ *
  * @author Patrick.Lau
  * @since 0.0.4 2023-10-03
  */
