+++ xcache-redis/xcache-redis-lettuce/src/main/java/com/igeeksky/redis/lettuce/LettuceOperator.java
@@ -3,14 +3,16 @@ package com.igeeksky.redis.lettuce;
 import io.lettuce.core.api.StatefulRedisConnection;
 
 /**
- * 单机 Redis 或 主从/哨兵 Redis 连接
+ * lettuce 客户端实现类
+ * <p>
+ * 支持（非集群）：standalone 及 sentinel 模式
  * <p>
  * 特别说明：无事务和阻塞命令，因此无需连接池
  *
  * @author Patrick.Lau
  * @since 0.0.4 2023-10-01
  */
-public class LettuceOperator extends AbstractLettuceOperator {
+public final class LettuceOperator extends AbstractLettuceOperator {
 
     public LettuceOperator(long timeout,
                            StatefulRedisConnection<byte[], byte[]> connection,
