+++ xcache-test/xcache-test-base/src/test/java/com/igeeksky/xcache/redis/lock/RedisSpinLockTest.java
@@ -1,6 +1,6 @@
 package com.igeeksky.xcache.redis.lock;
 
-import com.igeeksky.redis.lettuce.LettuceFactory;
+import com.igeeksky.redis.lettuce.LettuceStandaloneFactory;
 import com.igeeksky.redis.lettuce.LettuceTestHelper;
 import com.igeeksky.xcache.extension.lock.LockConfig;
 import com.igeeksky.xcache.extension.lock.LockTestTask;
@@ -22,7 +22,7 @@ import java.util.concurrent.locks.Condition;
 class RedisSpinLockTest {
 
     private static final String key = "test-lock";
-    private static LettuceFactory lettuceFactory;
+    private static LettuceStandaloneFactory lettuceStandaloneFactory;
     private static RedisLockService lockService;
 
     @BeforeAll
@@ -37,18 +37,18 @@ class RedisSpinLockTest {
                 .leaseTime(10000)
                 .build();
 
-        lettuceFactory = LettuceTestHelper.createStandaloneFactory();
+        lettuceStandaloneFactory = LettuceTestHelper.createStandaloneFactory();
 
         ScheduledExecutorService scheduler = Executors.newSingleThreadScheduledExecutor();
 
-        RedisLockProvider provider = new RedisLockProvider(scheduler, lettuceFactory);
+        RedisLockProvider provider = new RedisLockProvider(scheduler, lettuceStandaloneFactory);
 
         lockService = provider.get(config);
     }
 
     @AfterAll
     static void afterAll() {
-        lettuceFactory.close();
+        lettuceStandaloneFactory.close();
     }
 
     @Test
